Remove file split in RIC
